+++ core/src/com/nuclearthinking/myheroagency/controller/button/menu/PlayListener.java
@@ -3,8 +3,8 @@ package com.nuclearthinking.myheroagency.controller.button.menu;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.nuclearthinking.myheroagency.controller.ScreenEnum;
-import com.nuclearthinking.myheroagency.controller.ScreenManager;
 import com.nuclearthinking.myheroagency.controller.button.AbstractButtonListener;
+import com.nuclearthinking.myheroagency.controller.manager.ScreenManager;
 import lombok.NonNull;
 
 /**
