+++ core/src/com/nuclearthinking/myheroagency/ui/hud/layer/PlayerLayer.java
@@ -3,7 +3,7 @@ package com.nuclearthinking.myheroagency.ui.hud.layer;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
-import com.nuclearthinking.myheroagency.controller.ObjectManager;
+import com.nuclearthinking.myheroagency.controller.manager.ObjectManager;
 import com.nuclearthinking.myheroagency.controller.observer.ObserverCon;
 import com.nuclearthinking.myheroagency.controller.observer.ObserverMen;
 import com.nuclearthinking.myheroagency.ui.UiFactory;
